feat(rest): populate account settings + fix server_tuning (item-store economy) per Ghidra contract
Client (ClientRest::OnAccountResponse/ParseServerTuningBlock) reads settings + server_tuning;
was emitting empty settings + a broken server_tuning (offer_period=now, expiration off by 1000x,
Basic as int). Now emits the settings toggles + a sane offer window (24h) with unix-seconds
expiration and float multipliers. Contracts already matched the Ghidra field map.

+++ ReCap.Server/Adapters/Rest/Contracts/Game/Models/ServerTuningContract.cs
@@ -15,7 +15,7 @@ public class ServerTuningContract {
     public int? ItemstoreCurrentExpiration { get; set; }
 
     [XmlElement(ElementName = "itemstore_cost_multiplier_basic")]
-    public int? ItemstoreCostMultiplierBasic { get; set; }
+    public double? ItemstoreCostMultiplierBasic { get; set; }
 
     [XmlElement(ElementName = "itemstore_cost_multiplier_uncommon")]
     public double? ItemstoreCostMultiplierUncommon { get; set; }
